refactor: correct typo in ciruit_breaker module name
